refactor: use settings store for display settings
