+++ src/types/global.d.ts
@@ -1,0 +1,1 @@
+export type Currency = 'sats' | 'btc'
